example/django_auth_test/requirements.txt
Django
djangorestframework
python-dotenv
ikon-auth>=0.1.5
